requirements.txt
pyqt5
pyqt5-tools
pytesseract
PyPDF2
pillow
fpdf
opencv-python
tesseract
reportlab
pymupdf
